Add possibility to override Rank field for Jira via app settings (workaround for now)
As Jira does strange things, the lexo rank seems not to be always in the same custom field. Now the custom field can be specified in the app settings and overriden if not the default is used. This shall later be set per Jira instance on configuration, but it's good enough for now like this for the beta users