Allow WorkTrackingSystemOptions to be optional
this change is needed to support Jira DataCenter/Server instances, as they use a different means of authentication.
Jira Cloud uses an API token that requires specification of a user/email. Server/DataCenter is using a PAT, which does not require a username. In order to distinguish the two methods, we can check whether the user was specified or not. Thus we must allow omitting the username for Jira.